get sortBy result without displaying
