+++ index.html
@@ -51,6 +51,7 @@
     </div>
     <script src="./sources/data.js"></script>
     <script src="./js/functions.js"></script>
+    <script type="module" src="./js/checkbox-dom.js"></script>
     <!-- <script src="./js/filters.js"></script> -->
     <script type="module" src="./js/app.js"></script>
   </body>
